pyproject.toml
[project]
name = "auto-tts"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "26awen", email = "26awen@gmail.com" }
]
dependencies = [
    "pyttsx3>=2.90",
    "openai>=1.38.0",
    "xlwings>=0.31.10",
    "moviepy>=1.0.3",
    "pillow>=10.4.0",
    "numpy>=2.0.1",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/auto_tts"]
